requirements.txt
# Chikungunya Early Warning System - Requirements
# Python 3.9+ recommended

# Data & Numeric
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyarrow>=14.0.0  # For parquet support

# Machine Learning - Baselines
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Bayesian Modeling
pymc>=5.10.0
arviz>=0.17.0
cmdstanpy>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0
scikit-plot>=0.3.7

# Geospatial (optional)
geopandas>=0.14.0
shapely>=2.0.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0

# Notebook support
jupyter>=1.0.0
ipykernel>=6.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.1.0

# Fuzzy string matching (for district name matching)
rapidfuzz>=3.5.0
